Fixed glitches.
Glitch 1) Some key presses where showing up double. This was because of an if statement that should of been an if/else statement.

Glitch 2) The code would error at multiple points if a device file was found for a device that doesn't exist on the box. This was solved largely by adding
a property to the Device class called 'isValid' and using it as a check at multiple points in the code.

I also added another example config file and I have tested to make sure that it works correctly.